Made basic/foundation for AI
Made the Crabmonster move, look for player, and made patrolpoints we can place around the map for it to move to.  Made the Sentry follow the player, and stop after a certain amount of time outside of vision. It should also be rotating to look for the player, but can't remember if I uncommented the code that rotates it.